Add files via upload
This CSV contains the file name and folder name for the images that have missing values from the Zooniverse pulled down. It was manually filled in by Shannon (summer fellow) and Nikki (contracted research associate). This CSV is read into the code and fills in the missing file and folder names.