Lat_and_Long_data.csv
Latitude and Longitude for each folder uploaded to WWK so far. With each new upload, the base name of the folder (i.e. the grid location and the date) needs to be added to the document with the lat and long coordinates. This csv will need to be downloaded and is used within the WWK Cleaned Code.